updates
I created the 2D RASTER image of a star to perform the transformations on. the file is saved as star.png

I started familiarizing myself with the semantics and syntax of working with numpy and pillow interchangably

the createImage.py file is where I created and saved the star image.